+++ src/parsers/function_symbol.rs
@@ -2,6 +2,7 @@
 
 use crate::parsers::{parse_name, ParseResult, Span};
 use crate::types::FunctionSymbol;
+use nom::combinator::map;
 
 /// Parses a function symbol, i.e. `<name>`.
 ///
@@ -20,8 +21,7 @@ use crate::types::FunctionSymbol;
 /// assert!(parse_function_symbol("-1").is_err());
 ///```
 pub fn parse_function_symbol<'a, T: Into<Span<'a>>>(input: T) -> ParseResult<'a, FunctionSymbol> {
-    let (remaining, name) = parse_name(input)?;
-    Ok((remaining, name.into()))
+    map(parse_name, FunctionSymbol::new)(input.into())
 }
 
 impl crate::parsers::Parser for FunctionSymbol {