+++ src/types/name.rs
@@ -4,6 +4,15 @@ use crate::types::{PrimitiveType, ToTyped, Type, Typed};
 use std::fmt::{Debug, Display, Formatter};
 use std::ops::Deref;
 
+#[cfg(feature = "interning")]
+use std::sync::{Arc, Mutex};
+
+#[cfg(feature = "interning")]
+lazy_static::lazy_static! {
+    /// Used in [`Name::new_string_interned`] to deduplicate string occurrences.
+    static ref STRING_INTERNING: Mutex<Vec<Arc<String>>> = Mutex::new(Vec::default());
+}
+
 /// A name.
 ///
 /// ## Usage
@@ -18,12 +27,22 @@ pub struct Name(NameVariant);
 
 #[derive(Clone, PartialEq, Eq, Hash)]
 enum NameVariant {
-    String(String),
+    String(InternedString),
     Static(&'static str),
 }
 
+#[cfg(feature = "interning")]
+type InternedString = Arc<String>;
+
+#[cfg(not(feature = "interning"))]
+type InternedString = String;
+
 impl Name {
     /// Constructs a new [`Name`] from a provided string.
+    ///
+    /// ## Interning
+    /// If the `interning` crate feature is enabled, strings passed to this
+    /// method will be deduplicated, reducing memory footprint.
     ///
     /// ## Arguments
     /// * `name` - The name to wrap.
@@ -35,12 +54,18 @@ impl Name {
         if let Some(str) = Self::map_to_static(name.as_ref()) {
             Self::new_static(str)
         } else {
-            Self(NameVariant::String(name.into()))
+            Self::new_string_interned(name)
         }
     }
 
     /// Like [`new`] but makes use of the fact that if the string provided
     /// is `'static`, the method can be `const`.
+    ///
+    /// ## Interning
+    /// Note that strings passed to this method are not themselves interned
+    /// even if the create feature `interning` is enabled, and they will be invisible
+    /// to other strings partaking in interning. To ensure that a name value
+    /// exists exactly once, use the (non-const) [`Name::new`] function instead.
     ///
     /// ## Arguments
     /// * `name` - The name to wrap.
@@ -58,6 +83,34 @@ impl Name {
         Self(NameVariant::Static(name))
     }
 
+    /// Takes the provided `name` and interns the string.
+    ///
+    /// This uses a simple binary search approach to identify the correct position of
+    /// the input in question and inserts the element if it wasn't found before.
+    #[cfg_attr(not(feature = "interning"), inline(always))]
+    fn new_string_interned<S: Into<String> + AsRef<str>>(name: S) -> Self {
+        #[cfg(feature = "interning")]
+        {
+            let mut guard = STRING_INTERNING.lock().expect("failed to obtain lock");
+            let name_ref = name.as_ref();
+            let pos = guard.binary_search_by(|name| name_ref.cmp(name.as_str()));
+            let pos = match pos {
+                Ok(pos) => pos,
+                Err(pos) => {
+                    guard.insert(pos, Arc::new(name.into()));
+                    pos
+                }
+            };
+
+            Self(NameVariant::String(guard[pos].clone()))
+        }
+
+        #[cfg(not(feature = "interning"))]
+        {
+            Self(NameVariant::String(name.into()))
+        }
+    }
+
     pub fn is_empty(&self) -> bool {
         self.0.is_empty()
     }
@@ -70,8 +123,8 @@ impl Name {
     /// Maps the provided to a well-known `'static` string if possible.
     fn map_to_static<'a>(value: &'a str) -> Option<&'static str> {
         match value {
-            well_known::OBJECT => Some(well_known::OBJECT),
-            well_known::NUMBER => Some(well_known::NUMBER),
+            "object" => Some(well_known::OBJECT),
+            "number" => Some(well_known::NUMBER),
             _ => None,
         }
     }
@@ -79,8 +132,8 @@ impl Name {
 
 /// Provides well-known names for string interning.
 mod well_known {
-    pub const OBJECT: &'static str = "object";
-    pub const NUMBER: &'static str = "number";
+    pub static OBJECT: &'static str = "object";
+    pub static NUMBER: &'static str = "number";
 }
 
 impl<T> From<T> for Name
@@ -164,7 +217,7 @@ impl Deref for NameVariant {
     fn deref(&self) -> &Self::Target {
         match self {
             NameVariant::String(str) => str.as_str(),
-            NameVariant::Static(str) => str,
+            NameVariant::Static(str) => *str,
         }
     }
 }
@@ -172,8 +225,17 @@ impl Deref for NameVariant {
 impl PartialEq<str> for NameVariant {
     fn eq(&self, other: &str) -> bool {
         match self {
-            NameVariant::String(str) => str.eq(other),
             NameVariant::Static(str) => (*str).eq(other),
+            NameVariant::String(str) => {
+                #[cfg(feature = "interning")]
+                {
+                    str.as_ref().eq(other)
+                }
+                #[cfg(not(feature = "interning"))]
+                {
+                    str.eq(other)
+                }
+            }
         }
     }
 }
@@ -190,5 +252,19 @@ impl Display for NameVariant {
             NameVariant::String(str) => write!(f, "{}", str),
             NameVariant::Static(str) => write!(f, "{}", str),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use nom_greedyerror::AsStr;
+
+    #[test]
+    fn map_to_static_works() {
+        let object = Name::map_to_static("object").expect("mapping works");
+        let number = Name::map_to_static("number").expect("mapping works");
+        assert!(std::ptr::eq(object.as_str(), well_known::OBJECT));
+        assert!(std::ptr::eq(number.as_str(), well_known::NUMBER));
     }
 }